requirements.txt
requests
beautifulsoup4
tweepy
google-generativeai



